changed controllers and routes to simpler format
