+++ server/src/contollers/user.controller.ts
@@ -2,37 +2,56 @@ import UserModel from "../mongodb/models/user";
 import { Request, Response } from "express";
 
 // creates a user if the id doesnt exist already
-export const userCreate = async (email: string) => {
-    const userFound = await UserModel.findOne({email});
+export const userCreate = async (req: Request, res: Response) => {
+    
+    try {
+        const email = req.body;
+        const userFound = await UserModel.findOne({email});
 
-    if (userFound) throw new Error("User already exists!");
+        if (userFound) res.status(200).json(userFound._id)
 
-    const newUser = await UserModel.create({
-        email
-    })
+        const newUser = await UserModel.create({
+            email
+        })
 
-    return newUser._id;
+        res.status(200).json(newUser._id)
+    } catch (error) {
+        res.status(500).json({message: "User already exists"});
+    }
+
 }
 
-
 // delete user
-export const userDelete = async (email: string) => {
-    const userFound = await UserModel.findOne({email});
+export const userDelete = async (req: Request, res: Response) => {
+    
+    try {
+        const email = req.body;
+        const userFound = await UserModel.findOne({email});
 
-    if (!userFound) throw new Error("User not found!");
+        if (!userFound) throw new Error("User not found!");
 
-    await UserModel.deleteOne({email});
+        await UserModel.deleteOne({email});
 
-    return {}
+        res.status(200).json({})
+    } catch (error) {
+        res.status(500).json({message: "User doesnt exists"});
+    }
 
 }
 
 // get user email 
-export const userGet = async (id: number) => {
-    const userFound = await UserModel.findOne({_id: id}).populate("allProperties");
+export const userGet = async (req: Request, res: Response) => {
 
-    if (!userFound) throw new Error("User not found!");
+    try {
+        const id = parseInt(req.params.id);
 
-    return userFound.email;
+        const userFound = await UserModel.findOne({_id: id}).populate("allProperties");
+
+        if (!userFound) throw new Error("User not found!");
+
+        res.status(200).json(userFound.email)
+    } catch (error) {
+        res.status(500).json({message: "User doesnt exists"});
+    }
 
 }
